requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
PyYAML==6.0.2
requests==2.32.3